WIP
git-svn-id: http://svn.cangote.org@8 0ece9b63-09fb-49fa-bef2-506c893f8797

+++ core/gnunet/filesharing/transfer/downloaditem.h
@@ -66,7 +66,7 @@ public:
     DownloadItem *getParent();
     void modified();
     qint64 getRemainingDownload();
-    float getCompletedPercentage();
+    int getCompletedPercentage();
     void setStopped();
     void setPaused();
     void setCompleted();
@@ -78,6 +78,8 @@ public:
     void setETA(int eta);
     int getDataAdded();
     void setDataAdded(int date);
+
+
 private:
     DownloadItem * m_parent;
     struct GNUNET_FS_DownloadContext *m_context;
